requirements.txt
Django==4.2.1
pg8000
djangorestframework==3.14.0
python-dotenv==0.21.0
requests==2.28.1
dj-database-url==1.2.0
django-pg8000
